requirements.txt
google-genai
yfinance
plotly
pandas
streamlit